refactor(report): streamline test reports and lifecycle diagnostics
