+++ apps/report/e2e/midscene-test-lifecycle.yaml
@@ -1,0 +1,16 @@
+target:
+  serve: ./dist
+  url: demo-midscene-test-lifecycle.html
+  viewportWidth: 1440
+  viewportHeight: 1200
+tasks:
+  - name: lifecycle failures remain visible outside case attempts
+    flow:
+      - aiAssert: The overview reports a failed run and shows Project browser disposal failed, Document fixture disposal failed, Case session disposal failed, and Database connection refused. Web regression is expanded and includes Blocked by document setup.
+      - aiTap: the Blocked by document setup case row
+      - aiAssert: The not-run case has no Attempt selector. Its Execution panel shows the failed fixture.connect step under Document setup, and the evidence panel shows Database connection refused and the input host fixture-database.
+      - aiTap: the Overview back link
+      - aiTap: the Details action on the Web regression project row
+      - aiAssert: The project page preserves all four lifecycle failure messages, including the blocked document's Database connection refused error.
+      - aiTap: the Inspect action for Verify cart summary
+      - aiAssert: The case detail shows Case session disposal failed even though cart.verify itself succeeded, and no Attempt selector is shown for this single attempt.
