+++ apps/report/scripts/generate-demo-report.mjs
@@ -2,16 +2,14 @@
  * Run real Midscene YAML tests to generate reports, then create demo files:
  * - demo.html: single report (passed case)
  * - demo-merged.html: merged report with both passed and failed cases
- * - demo-midscene-test.html: unified Midscene Test hierarchy plus Agent details
  *
  * Usage: node scripts/generate-demo-report.mjs
- * Reuse an existing Agent report while iterating on the UI:
- * MIDSCENE_DEMO_SOURCE_REPORT=/absolute/report.html node scripts/generate-demo-report.mjs
  */
 import { execFileSync } from 'node:child_process';
 import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { generateTestReportFixtures } from '../e2e/fixtures/test-report.mjs';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const rootDir = path.join(__dirname, '..');
@@ -110,17 +108,7 @@ function copyReportWithUniqueGroupId(sourcePath, targetPath, suffix) {
 
 console.log('=== Generating passed report ===');
 const passedYaml = path.join(rootDir, 'scripts', 'generate-report.yaml');
-const requestedSourceReport = process.env.MIDSCENE_DEMO_SOURCE_REPORT;
-const passedReport = requestedSourceReport
-  ? path.resolve(repoRoot, requestedSourceReport)
-  : runYamlAndFindReport(passedYaml);
-if (requestedSourceReport && !fs.existsSync(passedReport)) {
-  console.error(`Demo source report does not exist: ${passedReport}`);
-  process.exit(1);
-}
-if (requestedSourceReport) {
-  console.log(`Reusing Agent report: ${passedReport}`);
-}
+const passedReport = runYamlAndFindReport(passedYaml);
 if (!passedReport) {
   console.error('Failed to generate passed report.');
   process.exit(1);
@@ -184,215 +172,10 @@ const demoMergedPath = path.join(distDir, 'demo-merged.html');
 fs.copyFileSync(mergedPath, demoMergedPath);
 console.log('Copied merged report -> dist/demo-merged.html');
 
-// --- Build a deterministic Midscene Test report around the same Agent dump ---
-const timestamp = '2026-08-20T08:00:00.000Z';
-const setupCompletedTimestamp = '2026-08-20T08:00:02.000Z';
-const firstAttemptCompletedTimestamp = '2026-08-20T08:00:03.000Z';
-const retryCompletedTimestamp = '2026-08-20T08:00:06.000Z';
-const testsCompletedTimestamp = '2026-08-20T08:00:07.000Z';
-const completedTimestamp = '2026-08-20T08:00:08.000Z';
-const baseStep = (id, node, status, stepIndex, extra = {}) => ({
-  id,
-  phase: 'steps',
-  stepIndex,
-  node,
-  status,
-  continuedAfterError: false,
-  startedAt: timestamp,
-  endedAt: completedTimestamp,
-  durationMs: 8000,
-  ...extra,
-});
-const testReportPath = new TestRunReportAssembler().assemble({
-  outputDir: distDir,
-  reportFileName: 'demo-midscene-test',
-  overwrite: true,
-  sources: [{ scopeId: 'attempt-2', sourcePath: passedReport }],
-  buildRunnerDump(index) {
-    const report = index.sources[0];
-    const executionId = report?.executionIds[0];
-    if (!executionId) {
-      throw new Error(
-        'The Midscene Test demo source has no stable execution ID.',
-      );
-    }
-    return {
-      schemaVersion: 1,
-      kind: 'test-runner',
-      runId: 'demo-run',
-      status: 'success',
-      startedAt: timestamp,
-      endedAt: completedTimestamp,
-      durationMs: 8000,
-      summary: {
-        total: 2,
-        passed: 2,
-        failed: 0,
-        notRun: 0,
-        filtered: 0,
-        collectionErrors: 0,
-        documentFailures: 0,
-        projectFailures: 0,
-      },
-      metrics: index.metrics,
-      projects: [
-        {
-          projectId: 'web-demo',
-          name: 'Web regression',
-          platform: 'web',
-          status: 'success',
-          retry: 1,
-          lifecycle: {
-            status: 'success',
-            startedAt: timestamp,
-            endedAt: completedTimestamp,
-            durationMs: 8000,
-          },
-          documents: [
-            {
-              documentId: 'checkout-document',
-              sourcePath: 'checkout.yaml',
-              status: 'success',
-              startedAt: setupCompletedTimestamp,
-              endedAt: testsCompletedTimestamp,
-              durationMs: 5000,
-              beforeAll: [
-                baseStep(
-                  'document-run:beforeAll:0',
-                  'fixture.prepare',
-                  'success',
-                  0,
-                  { phase: 'beforeAll', output: { summary: 'Fixture ready' } },
-                ),
-              ],
-              cases: [
-                {
-                  caseId: 'checkout-case',
-                  name: 'Checkout with retry',
-                  caseIndex: 0,
-                  status: 'success',
-                  attempts: [
-                    {
-                      attemptId: 'attempt-1',
-                      attemptIndex: 0,
-                      status: 'failed',
-                      startedAt: setupCompletedTimestamp,
-                      endedAt: firstAttemptCompletedTimestamp,
-                      durationMs: 1000,
-                      beforeEach: [],
-                      steps: [
-                        baseStep(
-                          'attempt-1:steps:0',
-                          'order.create',
-                          'failed',
-                          0,
-                          {
-                            durationMs: 20,
-                            input: {
-                              value: {
-                                sku: 'demo-item',
-                                apiKey: '[REDACTED]',
-                              },
-                              redactedPaths: ['$.apiKey'],
-                            },
-                            error: {
-                              name: 'NodeExecutionError',
-                              message: 'Inventory was not ready.',
-                              code: 'NODE_EXECUTION_ERROR',
-                            },
-                          },
-                        ),
-                      ],
-                      afterEach: [],
-                    },
-                    {
-                      attemptId: 'attempt-2',
-                      attemptIndex: 1,
-                      status: 'success',
-                      startedAt: firstAttemptCompletedTimestamp,
-                      endedAt: retryCompletedTimestamp,
-                      durationMs: 3000,
-                      beforeEach: [
-                        baseStep(
-                          'attempt-2:beforeEach:0',
-                          'browser.open',
-                          'success',
-                          0,
-                          {
-                            phase: 'beforeEach',
-                            durationMs: 10,
-                            output: { summary: 'Browser opened' },
-                          },
-                        ),
-                      ],
-                      steps: [
-                        baseStep('attempt-2:steps:0', 'aiAct', 'success', 0, {
-                          title:
-                            'Complete the checkout flow by selecting the saved address and submitting the demo order',
-                          agentDetails: [
-                            { reportId: report.reportId, executionId },
-                          ],
-                        }),
-                      ],
-                      afterEach: [
-                        baseStep(
-                          'attempt-2:afterEach:0',
-                          'order.cleanup',
-                          'success',
-                          0,
-                          {
-                            phase: 'afterEach',
-                            durationMs: 10,
-                            output: { summary: 'Order cleaned up' },
-                          },
-                        ),
-                      ],
-                      scopeReportIds: [report.reportId],
-                    },
-                  ],
-                },
-                {
-                  caseId: 'first-pass-case',
-                  name: 'Verify cart summary',
-                  caseIndex: 1,
-                  status: 'success',
-                  attempts: [
-                    {
-                      attemptId: 'attempt-3',
-                      attemptIndex: 0,
-                      status: 'success',
-                      startedAt: retryCompletedTimestamp,
-                      endedAt: testsCompletedTimestamp,
-                      durationMs: 1000,
-                      beforeEach: [],
-                      steps: [
-                        baseStep(
-                          'attempt-3:steps:0',
-                          'cart.verify',
-                          'success',
-                          0,
-                          {
-                            startedAt: retryCompletedTimestamp,
-                            endedAt: testsCompletedTimestamp,
-                            durationMs: 1000,
-                            output: {
-                              summary: 'Cart total and item count are correct',
-                            },
-                          },
-                        ),
-                      ],
-                      afterEach: [],
-                    },
-                  ],
-                },
-              ],
-              afterAll: [],
-            },
-          ],
-          collectionErrors: [],
-        },
-      ],
-    };
-  },
-});
-console.log(`Created Midscene Test demo -> ${testReportPath}`);
+for (const file of generateTestReportFixtures(
+  TestRunReportAssembler,
+  passedReport,
+  distDir,
+)) {
+  console.log(`Created report E2E fixture -> ${file}`);
+}
