+++ apps/report/src/components/test-runner/project.less
@@ -131,10 +131,6 @@
     font-variant-numeric: tabular-nums;
   }
 
-  .runner-detail-evidence-heading {
-    flex-wrap: wrap;
-  }
-
   .runner-detail-evidence-heading > div:first-child {
     min-width: 0;
     flex: 1 1 220px;
@@ -143,7 +139,6 @@
 
   .runner-detail-evidence-actions {
     min-width: 0;
-    flex-wrap: wrap;
     flex-shrink: 1;
   }
 
@@ -231,10 +226,6 @@
 
     .runner-header-actions {
       display: flex;
-    }
-
-    .runner-header-result {
-      display: none;
     }
 
     .runner-detail-trace-actions {
